tests/wpt/tests/shared-storage/cross-origin-create-worklet-failure-false-shared-storage-cross-origin-worklet-allowed.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>
<script src="/fenced-frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async t => {
  const ancestor_key = token();
  const crossOrigin = 'https://{{domains[www]}}:{{ports[https][0]}}';
  const helper_url = crossOrigin +
                     `/shared-storage/resources/credentials-test-helper.py` +
                     `?access_control_allow_origin_header=${window.origin}` +
                     `&access_control_allow_credentials_header=true` +
                     `&shared_storage_cross_origin_worklet_allowed_header=?0` +
                     `&token=${ancestor_key}`;

  return promise_rejects_dom(t, "OperationError",
    sharedStorage.createWorklet(
      helper_url + `&action=store-cookie`,
      { credentials: "include" }));
}, 'createWorklet() with cross-origin module script and credentials ' +
   '"include", and with the Shared-Storage-Cross-Origin-Worklet-Allowed ' +
   'response header value set to false (?0)');

</script>
</body>
